+++ thonny/running.py
@@ -39,7 +39,6 @@ from thonny.common import (
     ToplevelCommand,
     ToplevelResponse,
     UserError,
-    UserSystemExit,
     is_same_path,
     parse_message,
     path_startswith,
@@ -97,6 +96,7 @@ class Runner:
         self._publishing_events = False
         self._polling_after_id = None
         self._postponed_commands = []  # type: List[CommandToBackend]
+        self._last_accepted_backend_command = None
 
     def start(self) -> None:
         global _console_allocated
@@ -269,6 +269,7 @@ class Runner:
         else:
             assert response is None
             get_workbench().event_generate("CommandAccepted", command=cmd)
+            self._last_accepted_backend_command = cmd
             if isinstance(cmd, InlineCommand):
                 self._proxy.running_inline_command = True
 
@@ -332,6 +333,14 @@ class Runner:
         get_shell().submit_magic_command(cd_cmd_line + exe_cmd_line)
 
     def execute_editor_content(self, command_name, args):
+        if command_name.lower() == "debug":
+            messagebox.showinfo(
+                tr("Information"),
+                tr("For debugging the program must be saved first."),
+                master=get_workbench(),
+            )
+            return
+
         get_shell().submit_magic_command(
             construct_cmd_line(
                 ["%" + command_name, "-c", EDITOR_CONTENT_TOKEN] + args, [EDITOR_CONTENT_TOKEN]
@@ -344,6 +353,9 @@ class Runner:
         current file/script and submit it to shell
         """
 
+        if not self._proxy:
+            return
+
         assert (
             command_name[0].isupper()
             or command_name == "run"
@@ -354,12 +366,27 @@ class Runner:
             self._proxy.interrupt()
 
             try:
-                self._wait_for_prompt(2)
-            except TimeoutError as e:
-                get_shell().print_error(
-                    "Could not interrupt current process. Please wait, try again or select Stop/Restart!"
-                )
-                return
+                self._wait_for_prompt(1)
+            except TimeoutError:
+                # turtle.mainloop, for example, is not easy to interrupt.
+                get_shell().print_error("Could not interrupt current process. ")
+                wait_instructions = "Please wait, try again or select Stop/Restart!"
+
+                if self._proxy.stop_restart_kills_user_program():
+                    get_shell().print_error("Forcing the program to stop.\n")
+                    self.cmd_stop_restart()
+                    try:
+                        self._wait_for_prompt(1)
+                    except TimeoutError:
+                        get_shell().print_error(
+                            "Could not force-stop the program. " + wait_instructions + "\n"
+                        )
+                else:
+                    # For some back-ends (e.g. BareMetalMicroPython) killing is not an option.
+                    # Besides, they may be configured to avoid refreshing the environment
+                    # before run (for performance reasons).
+                    get_shell().print_error(wait_instructions + "\n")
+                    return
 
         editor = get_workbench().get_editor_notebook().get_current_editor()
         if not editor:
@@ -408,7 +435,7 @@ class Runner:
             get_workbench().update()
             sleep(0.01)
 
-        raise TimeoutError("Backend was not restarted in time")
+        raise TimeoutError("Backend did not respond in time")
 
     def _get_active_arguments(self):
         if get_workbench().get_option("view.show_program_arguments"):
@@ -508,7 +535,7 @@ class Runner:
         if get_workbench().in_simple_mode():
             get_workbench().hide_view("VariablesView")
 
-        self.restart_backend(True)
+        self.restart_backend(clean=True)
 
     def disconnect(self):
         proxy = self.get_backend_proxy()
@@ -567,15 +594,8 @@ class Runner:
 
                 msg_count += 1
             except BackendTerminatedError as exc:
-                self._handle_backend_termination(
-                    exc.returncode, getattr(self._proxy, "user_exit", False)
-                )
+                self._handle_backend_termination(exc.returncode)
                 return False
-
-            # TODO: temp hack
-            # Remember there was user exit, before the exit actually arrives
-            if isinstance(msg, UserSystemExit):
-                self._proxy.user_exit = True
 
             # change state
             if isinstance(msg, ToplevelResponse):
@@ -606,7 +626,7 @@ class Runner:
 
         self._send_postponed_commands()
 
-    def _handle_backend_termination(self, returncode: int, user_exit: bool) -> None:
+    def _handle_backend_termination(self, returncode: int) -> None:
         err = f"Process ended with exit code {returncode}."
 
         try:
@@ -621,8 +641,14 @@ class Runner:
 
         get_workbench().become_active_window(False)
         self.destroy_backend()
-        if user_exit:
+        if self._should_restart_after_command(self._last_accepted_backend_command):
             self.restart_backend(clean=True, automatic=True)
+
+    def _should_restart_after_command(self, cmd: CommandToBackend):
+        if cmd is None or not hasattr(cmd, "name"):
+            return False
+
+        return cmd.name in ["Run", "run", "Debug", "debug", "execute_source"]
 
     def restart_backend(self, clean: bool, first: bool = False, automatic: bool = False) -> None:
         """Recreate (or replace) backend proxy / backend process."""
@@ -823,6 +849,9 @@ class BackendProxy(ABC):
     def should_restart_interpreter_before_run(self) -> bool:
         return True
 
+    def stop_restart_kills_user_program(self) -> bool:
+        return True
+
     def supports_remote_directories(self):
         return False
 
@@ -891,7 +920,6 @@ class SubprocessProxy(BackendProxy, ABC):
         self._welcome_text = ""
 
         self._proc = None
-        self._terminated_readers = 0
         self._response_queue = None
         self._sys_path = []
         self._usersitepackages = None
@@ -993,12 +1021,9 @@ class SubprocessProxy(BackendProxy, ABC):
 
         logger.info("Starting the backend: %s %s", cmd_line, get_workbench().get_local_cwd())
 
-        extra_params = {}
-        if sys.version_info >= (3, 6):
-            extra_params["encoding"] = "utf-8"
-
         self._proc = subprocess.Popen(
             cmd_line,
+            executable=cmd_line[0],
             bufsize=0,
             stdin=subprocess.PIPE,
             stdout=subprocess.PIPE,
@@ -1007,13 +1032,12 @@ class SubprocessProxy(BackendProxy, ABC):
             env=self._get_environment(),
             universal_newlines=True,
             creationflags=creationflags,
-            **extra_params,
+            encoding="utf-8",
         )
 
         self._send_initial_input()
 
         # setup asynchronous output listeners
-        self._terminated_readers = 0
         Thread(target=self._listen_stdout, args=(self._proc.stdout,), daemon=True).start()
         Thread(target=self._listen_stderr, args=(self._proc.stderr,), daemon=True).start()
 
@@ -1030,6 +1054,7 @@ class SubprocessProxy(BackendProxy, ABC):
         """Send the command to backend. Return None, 'discard' or 'postpone'"""
         if isinstance(cmd, ToplevelCommand) and cmd.name[0].isupper():
             self._prepare_clean_launch()
+            logger.info("Prepared clean state for executing %r", cmd)
 
         if isinstance(cmd, ToplevelCommand):
             # required by SshCPythonBackend for creating fresh target process
@@ -1073,13 +1098,13 @@ class SubprocessProxy(BackendProxy, ABC):
 
     def _close_backend(self):
         if self._proc is not None and self._proc.poll() is None:
+            logger.info("Killing backend process")
             self._proc.kill()
 
         self._proc = None
         self._response_queue = None
 
     def _listen_stdout(self, stdout):
-        # debug("... started listening to stdout")
         # will be called from separate thread
 
         # allow self._response_queue to be replaced while processing
@@ -1137,10 +1162,7 @@ class SubprocessProxy(BackendProxy, ABC):
                                 )
                             )
 
-        self._terminated_readers += 1
-
     def _listen_stderr(self, stderr):
-        # stderr is used only for debugger debugging
         while True:
             data = read_one_incoming_message_str(stderr.readline)
             if data == "":
@@ -1149,8 +1171,6 @@ class SubprocessProxy(BackendProxy, ABC):
                 self._response_queue.append(
                     BackendEvent("ProgramOutput", stream_name="stderr", data=data)
                 )
-
-        self._terminated_readers += 1
 
     def _store_state_info(self, msg):
         if "cwd" in msg:
@@ -1203,7 +1223,7 @@ class SubprocessProxy(BackendProxy, ABC):
 
     def fetch_next_message(self):
         if not self._response_queue or len(self._response_queue) == 0:
-            if self.is_terminated() and self._terminated_readers == 2:
+            if self.is_terminated():
                 raise BackendTerminatedError(self._proc.returncode if self._proc else None)
             else:
                 return None