+++ thonny/tktextext.py
@@ -2,8 +2,9 @@
 """Extensions for tk.Text"""
 import sys
 import time
+import tkinter
 import tkinter as tk
-from logging import exception, getLogger
+from logging import getLogger
 from tkinter import TclError
 from tkinter import font as tkfont
 from tkinter import ttk
@@ -65,12 +66,23 @@ class TweakableText(tk.Text):
             ):
                 pass
             else:
-                exception(
+                logger.exception(
                     "[_dispatch_tk_operation] operation: " + operation + ", args:" + repr(args)
                 )
                 # traceback.print_exc()
 
             return ""  # Taken from idlelib.WidgetRedirector
+
+        except Exception as e:
+            # Need to catch to avoid the crash?
+            logger.exception("Exception in _dispatch_tk_operation")
+            from tkinter import messagebox
+
+            messagebox.showerror(
+                "Internal error",
+                "Error in _dispatch_tk_operation\n" + str(e),
+                parent=tkinter._default_root,
+            )
 
     def set_read_only(self, value):
         self._read_only = value
@@ -807,25 +819,27 @@ class TextFrame(ttk.Frame):
         }
         final_text_options.update(text_options)
         self.text = text_class(self, **final_text_options)
-        self.text.grid(row=0, column=2, sticky=tk.NSEW)
 
         if vertical_scrollbar:
             self._vbar = vertical_scrollbar_class(
                 self, orient=tk.VERTICAL, style=vertical_scrollbar_style
             )
-            self._vbar.grid(row=0, column=3, sticky=tk.NSEW)
             self._vbar["command"] = self._vertical_scroll
             self.text["yscrollcommand"] = self._vertical_scrollbar_update
+        else:
+            self._vbar = None
 
         if horizontal_scrollbar:
             self._hbar = horizontal_scrollbar_class(
                 self, orient=tk.HORIZONTAL, style=horizontal_scrollbar_style
             )
-            self._hbar.grid(row=1, column=0, sticky=tk.NSEW, columnspan=3)
             self._hbar["command"] = self._horizontal_scroll
             self.text["xscrollcommand"] = self._horizontal_scrollbar_update
+        else:
+            self._hbar = None
 
-        self.columnconfigure(2, weight=1)
+        self.grid_main_widgets()
+        self.columnconfigure(1, weight=1)
         self.rowconfigure(0, weight=1)
 
         self._ui_theme_change_binding = self.bind(
@@ -835,6 +849,13 @@ class TextFrame(ttk.Frame):
         # TODO: add context menu?
 
         self._reload_theme_options(None)
+
+    def grid_main_widgets(self):
+        self.text.grid(row=0, column=1, sticky=tk.NSEW)
+        if self._vbar:
+            self._vbar.grid(row=0, column=2, sticky=tk.NSEW)
+        if self._hbar:
+            self._hbar.grid(row=1, column=0, sticky=tk.NSEW, columnspan=3)
 
     def focus_set(self):
         self.text.focus_set()
@@ -1104,7 +1125,7 @@ class EnhancedTextFrame(TextFrame):
             ):  # In Python 3.6 you can use tk.EventType.ButtonPress instead of "4"
                 self.text.tag_remove("sel", "1.0", "end")
         except tk.TclError:
-            exception("on_gutter_click")
+            logger.exception("on_gutter_click")
 
     def on_gutter_double_click(self, event=None):
         try:
@@ -1112,7 +1133,7 @@ class EnhancedTextFrame(TextFrame):
             self.text.tag_remove("sel", "1.0", "end")
             self._gutter.tag_remove("sel", "1.0", "end")
         except tk.TclError:
-            exception("on_gutter_click")
+            logger.exception("on_gutter_click")
 
     def on_gutter_motion(self, event=None):
         try:
@@ -1126,7 +1147,7 @@ class EnhancedTextFrame(TextFrame):
             self.text.mark_set("insert", "%s.0" % linepos)
             self.text.focus_set()
         except tk.TclError:
-            exception("on_gutter_motion")
+            logger.exception("on_gutter_motion")
 
     def _vertical_scrollbar_update(self, *args):
         if not hasattr(self, "_vbar"):