+++ thonny/workbench.py
@@ -227,6 +227,14 @@ class Workbench(tk.Tk):
         self.ready = True
         self.event_generate("WorkbenchReady")
         self._editor_notebook.update_appearance()
+        if self._configuration_manager.error_reading_existing_file:
+            messagebox.showerror(
+                "Problem",
+                f"Previous configuration could not be read:\n\n"
+                f"{self._configuration_manager.error_reading_existing_file}).\n\n"
+                "Using default settings",
+                master=self,
+            )
 
     def _make_sanity_checks(self):
         home_dir = os.path.expanduser("~")
@@ -1715,9 +1723,12 @@ class Workbench(tk.Tk):
             return False
 
         if view.hidden:  # type: ignore
-            notebook.insert(
-                "auto", view.home_widget, text=self._view_records[view_id]["label"]  # type: ignore
-            )
+            label = None
+            if hasattr(view, "get_tab_text"):
+                label = view.get_tab_text()
+            if not label:
+                label = self._view_records[view_id]["label"]
+            notebook.insert("auto", view.home_widget, text=label)  # type: ignore
             view.hidden = False  # type: ignore
             if hasattr(view, "on_show"):  # type: ignore
                 view.on_show()
@@ -2358,7 +2369,7 @@ class Workbench(tk.Tk):
             self._closing = True
 
             runner = get_runner()
-            if runner != None:
+            if runner is not None:
                 runner.destroy_backend()
 
             # Tk clipboard gets cleared on exit and won't end up in system clipboard
@@ -2371,7 +2382,7 @@ class Workbench(tk.Tk):
                 ):
                     # Looks like the clipboard contains file name(s)
                     # Most likely this means actual file cut/copy operation
-                    # was made outside of Thonny.
+                    # was made outside Thonny.
                     # Don't want to replace this with simple string data of file names.
                     pass
                 else:
@@ -2404,6 +2415,8 @@ class Workbench(tk.Tk):
         sys.last_traceback = tb
         if isinstance(val, KeyboardInterrupt):
             # no need to report this, just let it close
+            logger.info("Got KeyboardInterrupt, closing")
+            self._on_close()
             return
         self.report_exception()